% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snowdata.R
\name{read_snowdata}
\alias{read_snowdata}
\title{Reading CSV File Sascha-Version}
\usage{
read_snowdata(file, start, value.na = NULL)
}
\arguments{
\item{file}{name of the file to read (format you sent me). Must
be character of length 1 (and exist).}

\item{start}{object of class POSIXt, length 1. Used as starting
point to generate the datetime variable.}

\item{value.na}{NULL or vector of the same class as the value column (your Z)
If not NULL, these values are replaced with NA.}
}
\value{

}
\description{
Reading the demo file you sent me and prepare some
things. I also assume a 'date' and use your time
information on X as 'hours after initialization'.
}
\examples{
x <- read_snowdata("data/datasascha.csv", as.POSIXct("2021-10-03"), value.na = 0)

}
\author{
Reto
}
