% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_snowdata_wfj.R
\name{read_snowdata_wfj}
\alias{read_snowdata_wfj}
\title{Reading Snowprofile Data from fwj ASCII File}
\usage{
read_snowdata_wfj(file, htol = 0.001)
}
\arguments{
\item{file}{name of the file to read; expects the file be
in the 'wfj' format you export.}

\item{htol}{numeric, height tolerance to double-check the
calculation, default is 0.001 (1mm). Must be larger
or equal to 0. See details.}
}
\value{
Returns a data.frame with with three variables named
}
\description{
Reading Snowprofile Data from fwj ASCII File
}
\details{
The columns \code{DZ_SNOW_M} contain the individual heights
of all snow layers. We modify the data set calculating the
top height of each layer by using \code{cumsum}. In addition
the data set contains \code{H_SNOW} which is the total height
of the snow pack.

We can then check the height above ground of the last snow
layer against \code{H_SNOW} which should be the same. Due to
the precision of the data in the input file there are small
differences. Thus, we use \code{htol} as a tolerance. If the
difference is smaller than \code{htol} we consider this
as a success, if larger deviations are found an error will
be thrown.
}
\examples{
x <- read_snowdata_wfj("data/202_snow_output_wfj.txt")

}
\author{
Reto
}
