% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdata_clean.R
\name{snowdata_clean}
\alias{snowdata_clean}
\title{Cleaning up Snowdata data set if needed}
\usage{
snowdata_clean(x, row.rm = FALSE)
}
\arguments{
\item{x}{the object returned by snowdata_to_zoo}

\item{row.rm}{logical, defaults to FALSE. If TRUe
rows not having any height or value (in combination)
will be removed. Will change the polygons which will
then bridge these gaps!}
}
\value{
Returns a cleaned \code{zoo} object nearly identical
to the input object \code{x} except that rows not containing
any meaningful data will be removed.

Note that if you use this function the polygons may be different
'briding' gaps. Maybe you don't want that anyways, wherefore this
function would be somewhat useless.
In addition the code is very similar to large parts of 
\code[snowprofi]{snowdata_to_sf}, if you keep it one should
try to streamline the two functions!
}
\description{
Removes rows (times) where no data is available. No data means
no height or no value for all levels.
}
\author{
Reto Stauffer
}
