% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdata_to_poly.R
\name{snowdata_create_one_poly}
\alias{snowdata_create_one_poly}
\title{Create single polygon}
\usage{
snowdata_create_one_poly(i)
}
\arguments{
\item{i}{integer, row-index.}
}
\description{
Used within \code{\link{snowdata_to_poly}}. Should not be used
directly.
}
\author{
Reto
}
