% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdata_to_poly.R
\name{snowdata_to_poly}
\alias{snowdata_to_poly}
\title{Convert Snowdata to Polygonformat}
\usage{
snowdata_to_poly(x)
}
\arguments{
\item{x}{object returned by \code[snowprofi]{snowdata_to_zoo} or
\code[snowprofi]{snowdata_clean}.}
}
\value{
An data.frame prepared to be used with \code[ggplot2]{geom_poly}.
}
\description{
Takes the object returned by \code[snowprofi]{snowdata_to_zoo}
and prepares a data.frame which can be used with \code[ggplot2]{geom_poly}.
The way it works is that you need a grouping identifier, and multiple (four)
points. The rest is then taken over by ggplot2.
}
\author{
Reto
}
