% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdata_to_sf.R
\name{snowdata_to_sf}
\alias{snowdata_to_sf}
\title{Convert Snowdata to an sf Object}
\usage{
snowdata_to_sf(x)
}
\arguments{
\item{x}{object returned by \code[snowprofi]{snowdata_to_zoo} or
\code[snowprofi]{snowdata_clean}.}
}
\value{
An sf object with polygon geometries and
\code{level} and \code{value} as variables for plotting.
}
\description{
This is the main function of the package. Taking some pre-prepared
\code{zoo} object returned by \code[snowprofi]{snowdata_to_zoo}
to create a bunch of polygons for each of the segments needed for
plotting.
}
\author{
Reto
}
