% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdata_to_zoo.R
\name{snowdata_to_zoo}
\alias{snowdata_to_zoo}
\title{Convert Snowdata to Zoo}
\usage{
snowdata_to_zoo(x, ...)
}
\arguments{
\item{x}{data.frame returned by \code[snowprofi]{read_snowdata}.}

\item{...}{currently ignored.}
}
\value{
Returns a multivariate zoo object with a distinct
way of naming the variables. This is required to be able
to rip it appart again later on when we convert the data
into a series of polygons for plotting.
}
\description{
Takes the object returned by \code[snowprofi]{read_snowdata}
and converts this object (data.frame) into a multivariate
zoo object. In addition, it performs a series of checks
}
\author{
Reto Stauffer
}
